% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{apibaseurl}
\alias{apibaseurl}
\title{Base HTTP Client functionality}
\usage{
apibaseurl()
}
\description{
\href{https://github.com/ropensci/seaaroundus/blob/682e7d8056e80160e14a25a3e94fbcffb650ed5c/R/helpers.R}{Pattern inspired by Scott Chamberlin at ROpenSci}
}
\details{
\dontrun{
callapi(path, args, ...)
}

\dontrun{
postapi(path, body, ...)
}
}
