% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_adult_counts.R
\name{daily_adult_counts}
\alias{daily_adult_counts}
\title{Get daily adult count data returned as a dataframe}
\usage{
daily_adult_counts(species = NULL, location = NULL)
}
\arguments{
\item{species}{(character) filter results by only one species.
Defaults to NULL, which will return all species}

\item{location}{(character) filter results by only one location
Defaults to NULL, which will return all locations}
}
\description{
Get daily adult count data returned as a dataframe
}
